refactor(business): rename src/code, src/error, src/model, and src/service to src/utils
This commit renames the src/code, src/error, src/model, and src/service directories to src/utils. This change is made to better reflect the purpose of these directories and to improve the organization of the codebase.

The src/utils directory now contains code that is used by multiple other directories in the project. This includes code for error handling, code generation, and data validation. By moving this code to a central location, it is easier to maintain and update.

This change does not affect the functionality of the project.
