+++ src/business/model/generic.model.ts
