+++ src/route/public.api.ts
@@ -1,5 +1,5 @@
 import express from "express";
-import { UserController } from "../controller/user.controller";
+import { UserController } from "../business/controller/user.controller";
 
 export const publicRouter = express.Router({
 	strict: true,
